+++ src/stream-statics.ts
@@ -3,8 +3,11 @@ import url from 'url'
 import fs from 'fs'
 import path from 'path'
 import { StstConfig, StringOrFalse } from './types'
-import mime from 'mime-types'
+import mime from './optional-mime.js'
 
+/*
+    All possible variables
+*/
 export const defaultConfig: StstConfig = {
     root: '.',
     port: 9630