+++ CHANGELOG.md
@@ -11,6 +11,9 @@
 
 ## [0.2.0]
 
+### Added
+- Opt-in to use mime-types package
+
 
 ## [0.1.0] - 2023-05-28